+++ frontend/src/components/home/MobileSuitTable.jsx
