+++ frontend/src/components/mobileSuits/parts/PartsInputList.jsx
@@ -1,0 +1,245 @@
+import React from "react";
+import { usePartsData } from "../../../hooks/mobileSuit/parts/inputList/usePartsData";
+import { usePartsOperations } from "../../../hooks/mobileSuit/parts/inputList/usePartsOperations";
+import { useMaterialSelection } from "../../../hooks/mobileSuit/parts/inputList/useMaterialSelection";
+
+export default function PartsInputList({
+  parts,
+  setParts,
+  msNumber,
+  partName,
+  functions,
+  setFunctions,
+  descriptions,
+  setDescriptions,
+  materials,
+  setMaterials
+}) {
+  // デバッグ用
+  console.log("PartsInputList props:", {
+    parts,
+    functions,
+    descriptions,
+    materials,
+    msNumber,
+    partName
+  });
+  // カスタムフックの使用（データ取得と初期化）
+  const {
+    initialPartsCount,
+    partTypeMap,
+  } = usePartsData(msNumber, partName, parts, setFunctions, setDescriptions, setMaterials);
+
+  // 親から渡された状態を直接使用
+  const currentFunctions = functions;
+  const currentDescriptions = descriptions;
+  const currentMaterials = materials;
+  
+  // デバッグ用
+  console.log("Current states:", {
+    currentFunctions,
+    currentDescriptions,
+    currentMaterials,
+    parts
+  });
+
+  const {
+    handleAdd,
+    handleRemove,
+    handlePartNameChange,
+    checkEditPermission,
+  } = usePartsOperations(
+    parts,
+    setParts,
+    msNumber,
+    partName,
+    initialPartsCount,
+    partTypeMap,
+    (partName) => {
+      setFunctions((prev) => {
+        const updated = { ...prev };
+        delete updated[partName];
+        return updated;
+      });
+      setDescriptions((prev) => {
+        const updated = { ...prev };
+        delete updated[partName];
+        return updated;
+      });
+      setMaterials((prev) => {
+        const updated = { ...prev };
+        delete updated[partName];
+        return updated;
+      });
+    }
+  );
+
+  const {
+    selectedIndex,
+    isDialogOpen,
+    materialOptions,
+    openMaterialDialog,
+    closeMaterialDialog,
+    toggleMaterial,
+    getSelectedMaterialsText,
+  } = useMaterialSelection(currentMaterials, setMaterials);
+
+  const handleChange = (index, field, value) => {
+    const partKey = parts[index];
+    if (!partKey?.trim()) return;
+
+    if (field === "function") {
+      setFunctions((prev) => ({ ...prev, [partKey]: value }));
+    } else if (field === "description") {
+      setDescriptions((prev) => ({ ...prev, [partKey]: value }));
+    } else if (field === "materials") {
+      setMaterials((prev) => ({ ...prev, [partKey]: value }));
+    }
+  };
+
+  const handlePartNameChangeWrapper = (index, newNameRaw) => {
+    const newName = newNameRaw.trim();
+    const oldName = parts[index];
+
+    const updatedParts = [...parts];
+    updatedParts[index] = newName;
+    setParts(updatedParts);
+
+    if (oldName === newName) return;
+
+    // 部品名変更時のデータ移行
+    setFunctions((prev) => {
+      const updated = { ...prev };
+      if (updated[oldName] !== undefined) {
+        updated[newName] = updated[oldName];
+        delete updated[oldName];
+      }
+      return updated;
+    });
+
+    setDescriptions((prev) => {
+      const updated = { ...prev };
+      if (updated[oldName] !== undefined) {
+        updated[newName] = updated[oldName];
+        delete updated[oldName];
+      }
+      return updated;
+    });
+
+    setMaterials((prev) => {
+      const updated = { ...prev };
+      if (updated[oldName] !== undefined) {
+        updated[newName] = updated[oldName];
+        delete updated[oldName];
+      }
+      return updated;
+    });
+  };
+
+  const handlePartNameClick = (index, partName) => {
+    checkEditPermission(index, partName);
+  };
+
+  return (
+    <div className="parts-input-group">
+      <h3>部品リスト</h3>
+      {parts.map((part, index) => (
+        <div key={index} className="part-input-row">
+          <label htmlFor={`part-name-${index}`}>部品名 {index + 1}</label>
+          <input
+            type="text"
+            value={part}
+            placeholder="部品名"
+            readOnly={index < initialPartsCount}
+            onClick={() => handlePartNameClick(index, part)}
+            onChange={(e) => {
+              if (index >= initialPartsCount) {
+                handlePartNameChangeWrapper(index, e.target.value);
+              }
+            }}
+          />
+
+          <div className="form-group">
+            <label htmlFor={`function-${index}`}>機能 {index + 1}</label>
+            <textarea
+              id={`function-${index}`}
+              name={`function-${index}`}
+              value={currentFunctions?.[part] ?? ""}
+              onChange={(e) => handleChange(index, "function", e.target.value)}
+            />
+          </div>
+
+          <div className="form-group">
+            <label htmlFor={`description-${index}`}>説明 {index + 1}</label>
+            <textarea
+              id={`description-${index}`}
+              name={`description-${index}`}
+              value={currentDescriptions?.[part] ?? ""}
+              onChange={(e) => handleChange(index, "description", e.target.value)}
+            />
+          </div>
+
+          <div className="form-group">
+            <button
+              type="button"
+              onClick={() => openMaterialDialog(index)}
+            >
+              素材選択
+            </button>
+            <span style={{ marginLeft: "10px" }}>
+              {getSelectedMaterialsText(part)}
+            </span>
+          </div>
+
+          <div className="form-group">
+            <label htmlFor={`partsImage-${index}`}>IMAGE FILE (.png)</label>
+            <input
+              type="file"
+              id={`partsImage-${index}`}
+              name="partsImage"
+              accept=".png"
+            />
+          </div>
+
+          <button type="button" onClick={() => handleRemove(index, currentFunctions, currentDescriptions)}>－ 削除</button>
+        </div>
+      ))}
+
+      <button type="button" onClick={handleAdd}>＋ 新規部品を追加</button>
+
+      {isDialogOpen && selectedIndex !== null && parts[selectedIndex] && (
+        <div className="modal-overlay" onClick={closeMaterialDialog}>
+          <div
+            className="modal-content"
+            onClick={(e) => e.stopPropagation()}
+            style={{
+              padding: "20px",
+              backgroundColor: "black",
+              borderRadius: "8px",
+              width: "300px",
+              margin: "100px auto",
+              color: "white",
+            }}
+          >
+            <h3>素材を選択してください</h3>
+            <ul style={{ listStyle: "none", padding: 0 }}>
+              {materialOptions.map((option) => (
+                <li key={option.code} style={{ marginBottom: "8px" }}>
+                  <label style={{ display: "block" }}>
+                    <input
+                      type="checkbox"
+                      checked={(currentMaterials[parts[selectedIndex]] ?? []).includes(option.label)}
+                      onChange={() => toggleMaterial(parts[selectedIndex], option.label)}
+                    />
+                    <span style={{ marginLeft: "8px" }}>{option.label}</span>
+                  </label>
+                </li>
+              ))}
+            </ul>
+            <button onClick={closeMaterialDialog}>完了</button>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+}
