+++ frontend/src/pages/login/Register.jsx
@@ -1,52 +1,19 @@
-import { useState } from "react";
-import { registerUser } from '../api/Controller'  
+import { useRegisterForm } from "../../hooks/login/useRegisterForm";
 
 export default function Register() {
-  const [formData, setFormData] = useState({
-    userId: "",
-    name: "",
-    password: "",
-    email: ""
-  });
-
-  const handleChange = (e) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({ ...prev, [name]: value }));
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    const { userId, name, password, email } = formData;
-
-    if (!userId || !name || !password || !email) {
-      alert("すべての項目を入力してください。");
-      return;
-    }
-
-    if (!/^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+$/.test(email)) {
-      alert("メールアドレスの形式が正しくありません");
-      return;
-    }
-
-    const confirmed = window.confirm("この情報で更新しますか？");
-    if (!confirmed) return;
-
-    try {
-	  const result = await registerUser(userId, name, password, email)
-	 
-	  alert(result.message || "登録完了！");
-	  
-	  setTimeout(() => {
+  const { formData, handleChange, submitForm } = useRegisterForm(
+    (msg) => {
+      alert(msg);
+      setTimeout(() => {
         window.parent.postMessage({ type: "registerCompleted" }, "*");
       }, 1500);
-	  
-	  } catch (error) {
-		
-      console.error('登録失敗:', error)
-      alert(result.message || "登録失敗")
-      
-      }
+    },
+    (errMsg) => alert(errMsg)
+  );
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    submitForm();
   };
 
   return (
@@ -55,22 +22,26 @@ export default function Register() {
       <form id="register-form" onSubmit={handleSubmit}>
         <div className="form-group" style={formGroupStyle}>
           <label htmlFor="userId" style={labelStyle}>社員ID</label>
-          <input type="text" id="userId" name="userId" value={formData.userId} onChange={handleChange} style={inputStyle} />
+          <input type="text" id="userId" name="userId"
+            value={formData.userId} onChange={handleChange} style={inputStyle} />
         </div>
 
         <div className="form-group" style={formGroupStyle}>
           <label htmlFor="name" style={labelStyle}>名前</label>
-          <input type="text" id="name" name="name" value={formData.name} onChange={handleChange} style={inputStyle} />
+          <input type="text" id="name" name="name"
+            value={formData.name} onChange={handleChange} style={inputStyle} />
         </div>
 
         <div className="form-group" style={formGroupStyle}>
           <label htmlFor="password" style={labelStyle}>パスワード</label>
-          <input type="password" id="password" name="password" value={formData.password} onChange={handleChange} style={inputStyle} />
+          <input type="password" id="password" name="password"
+            value={formData.password} onChange={handleChange} style={inputStyle} />
         </div>
 
         <div className="form-group" style={formGroupStyle}>
           <label htmlFor="email" style={labelStyle}>メールアドレス</label>
-          <input type="text" id="email" name="email" value={formData.email} onChange={handleChange} style={inputStyle} />
+          <input type="text" id="email" name="email"
+            value={formData.email} onChange={handleChange} style={inputStyle} />
         </div>
 
         <div className="button-group" style={buttonGroupStyle}>
@@ -82,7 +53,7 @@ export default function Register() {
 }
 
 const containerStyle = {
-  maxWidth: "600px",
+  maxWidth: "450px",
   margin: "auto",
   background: "#101010",
   border: "4px solid #333",
