+++ frontend/src/styles/login/Modal.css
@@ -1,0 +1,44 @@
+@charset "UTF-8";
+
+.modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.6);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+.modal-content {
+  background: #000000;
+  color: #00ffff;
+  padding: 25px;
+  border-radius: 8px;
+  min-width: 400px;
+  max-width: 600px;
+  position: relative;
+  box-shadow: 0 0 15px #00ffff88;
+  font-family: 'Meiryo', sans-serif;
+  text-shadow:
+    0 0 5px #00ffff,
+    0 0 10px #00ffff,
+}
+
+.modal-close-button {
+  position: absolute;
+  top: 10px;
+  right: 15px;
+  background: none;
+  border: none;
+  color: #00ffff;
+  font-size: 20px;
+  cursor: pointer;
+  text-shadow:
+    0 0 5px #00ffff,
+    0 0 10px #00ffff,
+    0 0 15px #00ffff;
+}
